Initialize android project
This change introduces the whole android project created by android
studio.

The change also includes a first functional version of the client. As of
now, it only sends a one time email with the subject of the search and
the email.

Next step is to offer to the client the possibility to register to
receive regular periodic mails about the subjects he is interested in.